Char LB UI
Setup for the leaderboard mostly complete. Has the sorting for the stats, and started defining split teams, descriptions, and preliminary stats. Now, need to fix the logic, as the visuals for sorting shows, but it doesn't actually sort or show on the backend yet. Also need to add the other characters.